requirements.txt
Flask>=3.0
requests>=2.32
pydantic>=2.8
pydantic-settings>=2.3
pydub>=0.25
Pillow>=10.0